+++ packages/svelte/tests/runtime-runes/samples/video-src-object/_config.js
@@ -1,0 +1,12 @@
+import { test } from '../../test';
+
+export default test({
+	html: `<video></video>`,
+
+	test({ assert, target }) {
+		const video = target.querySelector('video');
+
+		// @ts-ignore
+		assert.deepEqual(video?.srcObject, {});
+	}
+});